Migrate benchmarks to vitest bench
Compile legacy implementations for a fair and realistic benchmark comparison.
Remove obsolete benchmarks.

ref #20
