+++ .eslintignore
@@ -1,3 +1,3 @@
-benchmarks/
+legacy/
 dist/
 docs/
